+++ credsmash/api/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-
-from .delete import delete_secret
-from .get import get_secret
-from .list import list_secrets
-from .prune import prune_secret
-from .put import put_secret